requirements.txt
gTTS
SpeechRecognition
PyAudio
pygame
python-dotenv